Moved provider accommodation Id nearer to data
