+++ HappyTravel.Edo.Data/AccommodationMappings/ProviderAccommodationId.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using HappyTravel.Edo.Common.Enums;
 using Newtonsoft.Json;
 
-namespace HappyTravel.Edo.Api.Models.Accommodations
+namespace HappyTravel.Edo.Data.AccommodationMappings
 {
     public readonly struct ProviderAccommodationId
     {
